add biei's configuration.nix file
this isn't being used by this repo yet but I want to reference some
things in it from the other machine I'm setting up so adding it here so
I can :D

yes, I know my wifi password is in there. I'm not particularly concerned
about that. it's not that secret.