make everything into wee modules
yes I know some of these are just "programs.foo.enable = true;" but
that's fine. They're nice and contained and happy and if I want to
delete them I delete the file and the import and it's done. and any
extraneous configuration I want to do is contained nicely in the file.
and it's all organized and happy just like my old dotfiles used to be

nix is pretty rad