enable rose-pine theme for btop
thought about just copying the theme file over and to be honest probably
should have, but I like playing with things so here it is.

because I didn't want to manage the entire btop/plugins directory
as this whole repo, and I couldn't figure out how to use a relative path
to the theme, I decided to just grab the one file, which is somehow
(yea, I really want to dig more into that) limited to just the one file
by ... adding a path to the fetchFromGithub return? I really need to
look into how this works haha. I'm actually thinking fetchFromGithub
just returns a path to the ... derivation? hmm