work around programs.awscli's inistent management of credentials file
I don't want to put my creds in a repo so I was just going to manually make the file right there next to the config file. BUT. home-manager already manages that (even though it's empty). So put a differently formatted file next to it and read it with cat

side note: this was annoying to track down because I wasn't aware that "credentials_process" was a feature of awscli and not something fancy with home-manager or nix.
so I went trying to find the iniFormat.generate, not realizing that iniFormat was just a local variable name for pkgs.formats.ini, which is in nixpkgs, but that's also fairly abstracted and DRYed so was kinda hard to interpret

anyways, I just yolo and made it fake to see what happened and when I saw the output in the credentials file it made a lot more sense. hah.

thanks for coming to my ted talk